Refactored code to not duplicate functions needed in base abstract class
