+++ blender_applications/__init__.py
@@ -3,5 +3,6 @@ This Source Code Form is subject to the terms of the Mozilla Public
 License, v. 2.0. If a copy of the MPL was not distributed with this
 file, You can obtain one at https://mozilla.org/MPL/2.0/.
 """
-
-from .Win32BlenderApplication import Win32BlenderApplication
+from .blender_application import BlenderApplication
+from .win32_blender_application import Win32BlenderApplication
+from .darwin_blender_application import DarwinBlenderApplication
